researchmind/requirements.txt
crewai
crewai-tools
langchain
python-dotenv
pydantic
langchain-groq
litellm
fastapi
uvicorn
pylatex
